+++ src/main/java/iot/jcypher/domainquery/ast/PredicateExpression.java
@@ -1,5 +1,5 @@
 /************************************************************************
- * Copyright (c) 2014-2015 IoT-Solutions e.U.
+ * Copyright (c) 2014-2016 IoT-Solutions e.U.
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -125,6 +125,7 @@ public class PredicateExpression implements IASTObject {
 
 	/*****************************************************************/
 	public enum Operator {
-		EQUALS, NOT, LT, GT, LTE, GTE, LIKE, IN, CONTAINS, IS_NULL
+		EQUALS, NOT, LT, GT, LTE, GTE, LIKE, IN, CONTAINS, IS_NULL,
+		STARTS_WITH, ENDS_WITH, CONTAINS_STRING
 	}
 }